+++ CHANGELOG.md
@@ -7,6 +7,8 @@
    return `nil` as `Comparable#==` will not rescue exceptions in the next release.
  - Fix `Currency` specs for `#exponent` and `#decimal_places` not making assertions.
  - Fix a couple of Ruby warnings found in specs.
+ - Use `Money#decimal_mark` when formatting with `rounded_infinite_precision` rule
+   set to `true`.
 
 ## 6.6.0
  - Fixed VariableExchange#exchange_with for big numbers.